refactor: separate navigator parts
