+++ lib/features/bottom_nav_bar/bottom_nav_bar.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import '../../theme/app_theme.dart';
-import '../navigator/nested_navigator.dart';
+import '../navigator/navigator/nested_navigator.dart';
 import 'bottom_nav_bar_controller.dart';
 import 'nav_bar_config.dart';
