+++ lib/features/home_view/widgets/buildings_section/buildings_section.dart
@@ -9,7 +9,7 @@ import '../../../../utils/where_non_null_iterable.dart';
 import '../../../../widgets/my_error_widget.dart';
 import '../../../../widgets/subsection_header.dart';
 import '../../../bottom_nav_bar/nav_bar_config.dart';
-import '../../../navigator/nested_navigator.dart';
+import '../../../navigator/navigator/nested_navigator.dart';
 import '../../../buildings_map/controllers.dart';
 import '../loading_widgets/scrollable_section_loading.dart';
 import '../paddings.dart';
