+++ lib/features/home_view/widgets/parking_section.dart
@@ -6,7 +6,7 @@ import '../../../utils/where_non_null_iterable.dart';
 import '../../../widgets/my_error_widget.dart';
 import '../../../widgets/subsection_header.dart';
 import '../../bottom_nav_bar/nav_bar_config.dart';
-import '../../navigator/nested_navigator.dart';
+import '../../navigator/navigator/nested_navigator.dart';
 import '../../iparking/controllers.dart';
 import '../../iparking/models/parking_model.dart';
 import '../../iparking/repositories/parkings_repo.dart';
