+++ lib/features/navigator/navigator/detail_view_navigator.dart
@@ -1,0 +1,15 @@
+import '../../bottom_nav_bar/nav_bar_config.dart';
+import '../page_routes/topwr_page_route.dart';
+import 'nested_navigator.dart';
+import 'tab_bar_navigator.dart';
+
+/// Extracted logic of navigating to detailed views
+extension DetailViewNavigator on NestedNavigator {
+  void navigateToStudyCircleDetails(String argument) {
+    changeTabBar(NavBarEnum.sciCircles);
+    navigatorKey.currentState?.pushNamed(
+      AppRoutes.studyCircleDetails,
+      arguments: argument,
+    );
+  }
+}
