+++ lib/features/navigator/selected_tab_observer.dart
@@ -6,9 +6,10 @@ import 'utils/extract_tabbar_arg.dart';
 class SelectedTabObserver extends RouteObserver<PageRoute<dynamic>> {
   void Function(NavBarEnum tab) onTabChanged;
 
+  /// Watches the changes of new navigation routes and calls onTabChanged
   SelectedTabObserver(this.onTabChanged);
 
-  void recordCurrentlySelectedTab(Route route) {
+  void _recordCurrentlySelectedTab(Route route) {
     final routesTabData = route.settings.selectedTab;
     if (routesTabData != null) {
       onTabChanged(routesTabData);
@@ -18,26 +19,26 @@ class SelectedTabObserver extends RouteObserver<PageRoute<dynamic>> {
   @override
   void didPush(Route route, Route? previousRoute) {
     super.didPush(route, previousRoute);
-    recordCurrentlySelectedTab(route);
+    _recordCurrentlySelectedTab(route);
   }
 
   @override
   void didRemove(Route route, Route? previousRoute) {
     super.didRemove(route, previousRoute);
-    recordCurrentlySelectedTab(route);
+    _recordCurrentlySelectedTab(route);
   }
 
   @override
   void didReplace({Route? newRoute, Route? oldRoute}) {
     super.didReplace(newRoute: newRoute, oldRoute: oldRoute);
     if (newRoute != null) {
-      recordCurrentlySelectedTab(newRoute);
+      _recordCurrentlySelectedTab(newRoute);
     }
   }
 
   @override
   void didPop(Route route, Route? previousRoute) {
     super.didPop(route, previousRoute);
-    recordCurrentlySelectedTab(route);
+    _recordCurrentlySelectedTab(route);
   }
 }
