+++ lib/features/navigator/utils/android_pop_bug_workaround.dart
@@ -3,7 +3,7 @@ import 'dart:io';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 
-import '../nested_navigator.dart';
+import '../navigator/nested_navigator.dart';
 
 extension AndroidPopBugWorkaround on NestedNavigator {
   static const platform = MethodChannel('topwr.app.android.channel');
